refactored out the fixtures
* share the fixtures within two tests
* correct dependencies between fixtures
* use threading for cordinating between background and foreground
  threads.
