+++ iot/api-client/mqtt_example/cloudiot_mqtt_example_test.py
@@ -15,204 +15,38 @@
 import os
 import sys
 import time
-import uuid
 
-import backoff
-from googleapiclient.errors import HttpError
-from google.cloud import pubsub
-from google.api_core.exceptions import AlreadyExists
-from google.api_core.exceptions import NotFound
 import pytest
 
 # Add manager for bootstrapping device registry / device for testing
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', 'manager')) # noqa
 import cloudiot_mqtt_example
 import manager
-
+from fixtures import test_topic # noqa
+from fixtures import test_registry_id # noqa
+from fixtures import test_device_id # noqa
+from fixtures import device_and_gateways # noqa
 
 cloud_region = 'us-central1'
-device_id_template = 'test-device-{}'
 ca_cert_path = 'resources/roots.pem'
-rsa_cert_path = 'resources/rsa_cert.pem'
 rsa_private_path = 'resources/rsa_private.pem'
-topic_id = 'test-device-events-{}'.format(uuid.uuid4())
-
 project_id = os.environ['GCLOUD_PROJECT']
 service_account_json = os.environ['GOOGLE_APPLICATION_CREDENTIALS']
-
-pubsub_topic = 'projects/{}/topics/{}'.format(project_id, topic_id)
-
-# This format is used in the `../manager.py::clean_up_registries()`.
-registry_id = 'test-registry-{}-{}'.format(uuid.uuid4(), int(time.time()))
 
 mqtt_bridge_hostname = 'mqtt.googleapis.com'
 mqtt_bridge_port = 443
 
 
-@pytest.fixture(scope='module')
-def test_topic():
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def create_topic():
-        try:
-            return manager.create_iot_topic(project_id, topic_id)
-        except AlreadyExists as e:
-            # We ignore this case.
-            print("The topic already exists, detail: {}".format(str(e)))
-
-    topic = create_topic()
-
-    yield topic
-
-    pubsub_client = pubsub.PublisherClient()
-    topic_path = pubsub_client.topic_path(project_id, topic_id)
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def delete_topic():
-        try:
-            pubsub_client.delete_topic(topic_path)
-        except NotFound as e:
-            # We ignore this case.
-            print("The topic doesn't exist: detail: {}".format(str(e)))
-
-    delete_topic()
-
-
-@pytest.fixture(scope='module')
-def test_registry_id():
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def create_registry():
-        manager.open_registry(
-            service_account_json, project_id, cloud_region, pubsub_topic,
-            registry_id)
-
-    create_registry()
-
-    yield registry_id
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def delete_registry():
-        try:
-            manager.delete_registry(
-                service_account_json, project_id, cloud_region, registry_id)
-        except NotFound as e:
-            # We ignore this case.
-            print("The registry doesn't exist: detail: {}".format(str(e)))
-
-    delete_registry()
-
-
-@pytest.fixture(scope='module')
-def rsa256_device_id(test_registry_id):
-    device_id = device_id_template.format('RSA256')
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def create_device():
-        manager.create_rs256_device(
-            service_account_json, project_id, cloud_region, test_registry_id,
-            device_id, rsa_cert_path)
-
-    create_device()
-
-    yield device_id
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def delete_device():
-        try:
-            manager.delete_device(
-                service_account_json, project_id, cloud_region,
-                test_registry_id, device_id)
-        except NotFound as e:
-            # We ignore this case.
-            print("The device doesn't exist: detail: {}".format(str(e)))
-
-    delete_device()
-
-
-@pytest.fixture(scope='module')
-def device_and_gateways():
-    device_id = device_id_template.format('noauthbind')
-    gateway_id = device_id_template.format('RS256')
-    bad_gateway_id = device_id_template.format('RS256-err')
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def create_device():
-        manager.create_device(
-            service_account_json, project_id, cloud_region, registry_id,
-            device_id)
-    create_device()
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def create_gateways():
-        manager.create_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            None, gateway_id, rsa_cert_path, 'RS256')
-        manager.create_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            None, bad_gateway_id, rsa_cert_path, 'RS256')
-
-    create_gateways()
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def bind_device_to_gateways():
-        manager.bind_device_to_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            device_id, gateway_id)
-        manager.bind_device_to_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            device_id, bad_gateway_id)
-
-    bind_device_to_gateways()
-
-    yield (device_id, gateway_id, bad_gateway_id)
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def unbind():
-        manager.unbind_device_from_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            device_id, gateway_id)
-        manager.unbind_device_from_gateway(
-            service_account_json, project_id, cloud_region, registry_id,
-            device_id, bad_gateway_id)
-
-    unbind()
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def delete_device():
-        try:
-            manager.delete_device(
-                service_account_json, project_id, cloud_region, registry_id,
-                device_id)
-        except NotFound as e:
-            # We ignore this case.
-            print("The device doesn't exist: detail: {}".format(str(e)))
-
-    delete_device()
-
-    @backoff.on_exception(backoff.expo, HttpError, max_time=60)
-    def delete_gateways():
-        try:
-            manager.delete_device(
-                service_account_json, project_id, cloud_region, registry_id,
-                gateway_id)
-            manager.delete_device(
-                service_account_json, project_id, cloud_region, registry_id,
-                bad_gateway_id)
-        except NotFound as e:
-            # We ignore this case.
-            print("The gateway doesn't exist: detail: {}".format(str(e)))
-
-    delete_gateways()
-
-
-def test_event(test_topic, test_registry_id, rsa256_device_id, capsys):
+def test_event(test_topic, test_registry_id, test_device_id, capsys): # noqa
     manager.get_device(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     sub_topic = 'events'
-    mqtt_topic = '/devices/{}/{}'.format(rsa256_device_id, sub_topic)
+    mqtt_topic = '/devices/{}/{}'.format(test_device_id, sub_topic)
 
     client = cloudiot_mqtt_example.get_client(
-        project_id, cloud_region, test_registry_id, rsa256_device_id,
+        project_id, cloud_region, test_registry_id, test_device_id,
         rsa_private_path, 'RS256', ca_cert_path,
         'mqtt.googleapis.com', 443)
 
@@ -223,22 +57,22 @@ def test_event(test_topic, test_registry_id, rsa256_device_id, capsys):
 
     manager.get_state(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     out, _ = capsys.readouterr()
     assert 'on_publish' in out
 
 
-def test_state(test_topic, test_registry_id, rsa256_device_id, capsys):
+def test_state(test_topic, test_registry_id, test_device_id, capsys): # noqa
     manager.get_device(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     sub_topic = 'state'
-    mqtt_topic = '/devices/{}/{}'.format(rsa256_device_id, sub_topic)
+    mqtt_topic = '/devices/{}/{}'.format(test_device_id, sub_topic)
 
     client = cloudiot_mqtt_example.get_client(
-        project_id, cloud_region, test_registry_id, rsa256_device_id,
+        project_id, cloud_region, test_registry_id, test_device_id,
         rsa_private_path, 'RS256', ca_cert_path,
         'mqtt.googleapis.com', 443)
     client.publish(mqtt_topic, 'state test', qos=1)
@@ -250,20 +84,20 @@ def test_state(test_topic, test_registry_id, rsa256_device_id, capsys):
 
     manager.get_state(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     out, _ = capsys.readouterr()
     assert 'on_publish' in out
     assert 'binary_data: "state test"' in out
 
 
-def test_config(test_topic, test_registry_id, rsa256_device_id, capsys):
+def test_config(test_topic, test_registry_id, test_device_id, capsys): # noqa
     manager.get_device(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     client = cloudiot_mqtt_example.get_client(
-        project_id, cloud_region, test_registry_id, rsa256_device_id,
+        project_id, cloud_region, test_registry_id, test_device_id,
         rsa_private_path, 'RS256', ca_cert_path,
         'mqtt.googleapis.com', 443)
     client.loop_start()
@@ -274,18 +108,18 @@ def test_config(test_topic, test_registry_id, rsa256_device_id, capsys):
 
     manager.get_state(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id)
+            test_device_id)
 
     out, _ = capsys.readouterr()
     assert "Received message" in out
-    assert '/devices/{}/config'.format(rsa256_device_id) in out
+    assert '/devices/{}/config'.format(test_device_id) in out
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1)
-def test_receive_command(test_registry_id, rsa256_device_id, capsys):
+def test_receive_command(test_registry_id, test_device_id, capsys): # noqa
     # Exercize the functionality
     client = cloudiot_mqtt_example.get_client(
-        project_id, cloud_region, test_registry_id, rsa256_device_id,
+        project_id, cloud_region, test_registry_id, test_device_id,
         rsa_private_path, 'RS256', ca_cert_path,
         'mqtt.googleapis.com', 443)
     client.loop_start()
@@ -297,7 +131,7 @@ def test_receive_command(test_registry_id, rsa256_device_id, capsys):
 
     manager.send_command(
             service_account_json, project_id, cloud_region, test_registry_id,
-            rsa256_device_id, 'me want cookies')
+            test_device_id, 'me want cookies')
 
     # Process commands
     for i in range(1, 5):
@@ -311,7 +145,7 @@ def test_receive_command(test_registry_id, rsa256_device_id, capsys):
 
 @pytest.mark.flaky(max_runs=5, min_passes=1)
 def test_gateway_listen_for_bound_device_configs(
-        test_topic, test_registry_id, device_and_gateways, capsys):
+        test_topic, test_registry_id, device_and_gateways, capsys): # noqa
     (device_id, gateway_id, _) = device_and_gateways
 
     # Setup for listening for config messages
@@ -332,7 +166,7 @@ def test_gateway_listen_for_bound_device_configs(
 
 @pytest.mark.flaky(max_runs=5, min_passes=1)
 def test_gateway_send_data_for_device(
-        test_topic, test_registry_id, device_and_gateways, capsys):
+        test_topic, test_registry_id, device_and_gateways, capsys): # noqa
     (device_id, gateway_id, _) = device_and_gateways
 
     # Setup for listening for config messages
@@ -342,10 +176,10 @@ def test_gateway_send_data_for_device(
 
     # Connect the gateway
     cloudiot_mqtt_example.send_data_from_bound_device(
-                service_account_json, project_id, cloud_region, registry_id,
-                device_id, gateway_id, num_messages, rsa_private_path,
-                'RS256', ca_cert_path, mqtt_bridge_hostname, mqtt_bridge_port,
-                jwt_exp_time, listen_time)
+        service_account_json, project_id, cloud_region, test_registry_id,
+        device_id, gateway_id, num_messages, rsa_private_path, 'RS256',
+        ca_cert_path, mqtt_bridge_hostname, mqtt_bridge_port, jwt_exp_time,
+        listen_time)
 
     out, _ = capsys.readouterr()
     assert 'Publishing message 5/5' in out
@@ -353,7 +187,7 @@ def test_gateway_send_data_for_device(
 
 
 def test_gateway_trigger_error_topic(
-        test_topic, test_registry_id, device_and_gateways, capsys):
+        test_topic, test_registry_id, device_and_gateways, capsys): # noqa
     (device_id, _, gateway_id) = device_and_gateways
 
     # Setup for listening for config messages
@@ -365,16 +199,14 @@ def test_gateway_trigger_error_topic(
 
     # Connect the gateway
     cloudiot_mqtt_example.listen_for_messages(
-                service_account_json, project_id, cloud_region, registry_id,
-                device_id, gateway_id, num_messages, rsa_private_path,
-                'RS256', ca_cert_path, 'mqtt.googleapis.com', 443,
-                20, 42, trigger_error)
+        service_account_json, project_id, cloud_region, test_registry_id,
+        device_id, gateway_id, num_messages, rsa_private_path, 'RS256',
+        ca_cert_path, 'mqtt.googleapis.com', 443, 20, 42, trigger_error)
     # Try to connect the gateway aagin on 8883
     cloudiot_mqtt_example.listen_for_messages(
-                service_account_json, project_id, cloud_region, registry_id,
-                device_id, gateway_id, num_messages, rsa_private_path,
-                'RS256', ca_cert_path, 'mqtt.googleapis.com', 8883,
-                20, 15, trigger_error)
+        service_account_json, project_id, cloud_region, test_registry_id,
+        device_id, gateway_id, num_messages, rsa_private_path, 'RS256',
+        ca_cert_path, 'mqtt.googleapis.com', 8883, 20, 15, trigger_error)
 
     out, _ = capsys.readouterr()
     assert 'GATEWAY_ATTACHMENT_ERROR' in out
